+++ example/mica-mqtt-client-spring-boot-example/src/main/java/net/dreamlu/iot/mqtt/client/service/ClientService.java
@@ -20,7 +20,7 @@ public class ClientService {
 	private MqttClientTemplate client;
 
 	public boolean publish(String body) {
-		client.publish("/test/client", ByteBuffer.wrap(body.getBytes(StandardCharsets.UTF_8)));
+		client.publish("/test/client", body.getBytes(StandardCharsets.UTF_8));
 		return true;
 	}